+++ src/Network/HTTP/Download/Verified.hs
@@ -9,6 +9,7 @@
 module Network.HTTP.Download.Verified
   ( verifiedDownload
   , DownloadRequest(..)
+  , drRetriesDefault
   , HashCheck(..)
   , CheckHexDigest(..)
   , LengthCheck
@@ -50,8 +51,13 @@ data DownloadRequest = DownloadRequest
     { drRequest :: Request
     , drHashChecks :: [HashCheck]
     , drLengthCheck :: Maybe LengthCheck
+    , drRetries :: Int
     }
   deriving Show
+
+-- | Default to retrying thrice.
+drRetriesDefault :: Int
+drRetriesDefault = 3
 
 data HashCheck = forall a. (Show a, HashAlgorithm a) => HashCheck
   { hashCheckAlgorithm :: a
@@ -84,6 +90,8 @@ data VerifiedDownloadException
           String -- algorithm
           CheckHexDigest -- expected
           String -- actual (shown)
+    | ZeroTries
+          Request
   deriving (Typeable)
 instance Show VerifiedDownloadException where
     show (WrongContentLength req expected actual) =
@@ -100,6 +108,10 @@ instance Show VerifiedDownloadException where
         "Download expectation failure: content hash (" ++ algo ++  ")\n"
         ++ "Expected: " ++ displayCheckHexDigest expected ++ "\n"
         ++ "Actual:   " ++ actual ++ "\n"
+        ++ "For: " ++ show (getUri req)
+    show (ZeroTries req) =
+        "Download expectation failure:\n"
+        ++ "Download was needed but <= 0 retries were requested.\n"
         ++ "For: " ++ show (getUri req)
 
 instance Exception VerifiedDownloadException
@@ -168,6 +180,24 @@ sinkHashUsing _ = sinkHash
 hashChecksToZipSink :: MonadThrow m => Request -> [HashCheck] -> ZipSink ByteString m ()
 hashChecksToZipSink req = traverse_ (ZipSink . sinkCheckHash req)
 
+-- TODO(DanBurton): use Control.Retry instead.
+-- Type inference drives the decision of which exceptions merit a retry.
+retry :: (MonadCatch m, Exception e)
+  => Int -- ^ The number of times to retry
+  -> m a -- ^ Action to retry
+  -> m (Either [e] a)
+retry n0 action =
+    go n0 []
+  where
+    go n es
+      | n <= 0 = return (Left es)
+      | otherwise = do
+          eRes <- try action
+          case eRes of
+            Left e -> go (n - 1) (e : es)
+            Right a -> return (Right a)
+
+
 -- | Copied and extended version of Network.HTTP.Download.download.
 --
 -- Has the following additional features:
@@ -178,7 +208,9 @@ hashChecksToZipSink req = traverse_ (ZipSink . sinkCheckHash req)
 -- * Verifies md5 if response includes content-md5 header
 -- * Verifies the expected hashes
 --
--- Throws VerifiedDownloadException, and whatever else "download" throws.
+-- Throws VerifiedDownloadException.
+-- Throws IOExceptions related to file system operations.
+-- Throws HttpException.
 verifiedDownload :: (MonadReader env m, HasHttpManager env, MonadIO m)
          => DownloadRequest
          -> Path Abs File -- ^ destination
@@ -189,9 +221,14 @@ verifiedDownload DownloadRequest{..} destpath progressSink = do
     env <- ask
     liftIO $ whenM' getShouldDownload $ do
         createDirectoryIfMissing True dir
-        withBinaryFile fptmp WriteMode $ \h ->
-            flip runReaderT env $
-                withResponse req (go h)
+        withBinaryFile fptmp WriteMode $ \h -> do
+            eRes <- retry drRetries $
+                flip runReaderT env $
+                    withResponse req (go h)
+            case (eRes :: Either [HttpException] ()) of
+                Left [] -> throwM $ ZeroTries req
+                Left (e:_) -> throwM e -- just re-throw the latest HttpException
+                Right () -> return ()
         renameFile fptmp fp
   where
     whenM' mp m = do